Deleted DANClass2, since DANClass was being used instead. Changed getNEUTRON_COUNT return type to int.

+++ src/DAN/server/DANClass.java
@@ -59,7 +59,7 @@ public class DANClass {
 		this.DE_TIME = date;
 	}
 
-	public double getNEUTRON_COUNT() {
+	public int getNEUTRON_COUNT() {
 		return NEUTRON_COUNT;
 	}
